# Improve performance - skip manifest read if no components/targets requested