I was looking into why rustup had a performance penalty over calling cargo directly in simple cases, and found most of the time was spent parsing the `multirust-channel-manifest.toml` file. Looking more closely, it seemed parsing that file was not really required in many cases. This PR tweaks the ordering of the code to skip parsing this file when it isn't needed. 

I do not intend to make any behavioral change with this, other than potentially if some code paths handle an error while reading this file (and now they won't see that error in cases where we skip reading the file entirely). I believe this is the reason the `heal_damaged_toolchain` test is now failing. I need to look into this more closely (this PR is marked as draft until I figure this out), but I'd appreciate any insight into whether that test was intended to cover this behavior. 

The performance changes were tested in a brand new hello world cargo project (`cargo new hello`), after building rustup in release mode and installing it into a `rustup/home` directory as described in the contributing guidelines. The before tests were done on the current master branch, because it seems to already be much faster than the most recent official release. 

Before this change a repeated call to `cargo run` (through the rustup proxy) would take ~90ms. After this change it takes ~35ms. Running cargo directly takes ~22ms. So this change **reduces the rustup overhead in this case from 68ms to 13ms (a reduction of 80%)**.