@JoshMcguigan 2021-12-04
It looks like the `heal_damaged_toolchain` test is failing because it relies on `rustup show active-toolchain` having previously read/parsed the manifest, which is specifically what this PR removes (since reading/parsing the manifest seems to not be strictly required for that operation, and it takes meaningful time). 

Is my understanding correct here? And is it reasonable to a command like `rustup show active-toolchain` to NOT heal the damaged toolchain manifest file (since it doesn't even need to look at that file)? If so, is there a command which would be more appropriate to run as part of this test to demonstrate the healing?