@kinnison 2021-12-27
Personally I think the request to show the active toolchain (which implies checking for components in the `rust-toolchain.toml` being available) should be able to succeed at healing things.

I remain convinced that the correct way to ameliorate the incredible slowness of parsing the TOML is to, instead, keep a cached faster-to-parse copy alongside it, e.g. BINCODE.