@dralley 2022-01-12
Has any profiling been done on the TOML parsing itself?  Just glancing through the file, the parsing code doesn't seem especially efficient, it could probably benefit from a little focused attention.