@JoshMcguigan 2022-01-12
I created a small one.

```rust
use std::fs::read_to_string;

fn main() {
    let s = read_to_string("/home/josh/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/multirust-channel-manifest.toml").unwrap();

    for _ in 0..10 {
        s.parse::<toml::Value>().unwrap();
    }
}
```

This runs in ~450ms on my machine. I found I can improve this by ~100ms by bypassing the `\r\n` windows style newline handling in `src/tokens.rs`, which I actually wonder if that is where you are seeing all the allocations. However I'm not sure how we could get this improvement while also handling windows style newlines (short of replacing all `\r\n` with `\n` before parsing, but there is a memory use cost there).