@dralley 2022-01-12
> I'd like to poke around at this. How did you gather this data?

If you pass the `--reverse` flag it groups together all of the tiny calls at the bottom of the call stack instead of the ones at the top, if you do it seems that about 25% of the total time is spent on allocs, frees and moves.