@JoshMcguigan 2022-01-13
I've been using frida tracing to count malloc calls, and its about 100,000 allocations to parse the manifest file a single time (so the example above would be 10x that). I replaced `Vec` in one place with `TinyVec` and cut the allocations by ~10%, but the runtime didn't really change. 

I'm probably not going to continue down this path right now, but I'd be very curious to see if anyone else comes up with anything fruitful here.